Feat: Entering text will render 'TEXT CHANGED' in results area.
a handleChange function is added to ContainerArea class and passed to TextArea via onChange props.
input property is added to this.state in ContainerArea by getInitialState.
the input property stores what would be displayed in the result area.
In handleChange method, the value of input is changed to 'TEXT CHANGED' by setState.
Next step will be displaying exatly what is being entered in th textarea.