Feat: Markup in text area is now converted using marked.
As the user types markdown in textarea, the text is directly formatted properly and displayed in results area.
Changes are mafe ResultArea component. A text prop is added ResultArea component within ContainerArea.
rawMarkup method added to ResultArea to convert Markdown.
The call to rawMarkup is occures from span with prop dangerouslySetInnerHTML.